Subsumes and Equivalent, for valditity expanding and preserving operations
