+++ typic/src/transmute/from_type.rs
@@ -1,7 +1,7 @@
 use super::from_layout;
 
-/// A marker trait implemented if every instance of `T` is transmutable into
-/// an instance of `Self`.
-pub unsafe trait FromType<T> {}
+mod subsumes;
+mod equivalent;
 
-unsafe impl<T, U> FromType<U> for T where U: from_layout::FromLayout<T> {}
+pub use subsumes::Subsumes;
+pub use equivalent::Equivalent;
